+++ ui/src/app/_services/api.service.ts
@@ -75,11 +75,25 @@ export class ApiService {
   }
 
   loadConfig() {
-  return this.$http.get(`${this.base}/api/config`, Object.assign({ responseType: 'text' as 'text' }, this.httpOptions));
+  return this.$http.get(`${this.base}/api/config`,
+    Object.assign({ responseType: 'text' as 'text' }, this.httpOptions));
   }
 
   saveConfig(config) {
     return this.$http.post(`${this.base}/api/config`, config, this.httpOptions);
+  }
+
+  getConfigBackupList() {
+    return this.$http.get(`${this.base}/api/config/backups`, this.httpOptions);
+  }
+
+  getConfigBackup(backupId) {
+    return this.$http.get(`${this.base}/api/config/backups/${backupId}`,
+      Object.assign({ responseType: 'text' as 'text' }, this.httpOptions));
+  }
+
+  deleteConfigBackups() {
+    return this.$http.delete(`${this.base}/api/config/backups`, this.httpOptions);
   }
 
   getUsers() {